GOOD SPOT
rnn_weatherData.ipynb. is now a GREAT working PyTorch RNN. it takes in the weather data, and breaks it into 3 sets: training, validation, testing. Where the training and validation sets are used to train the inital RNN. once the rnn is trained, you then run a single pass through that rnn in evaluation mode .eval() , it then makes a predition for that test data set off only a single pass through. this test pass does not update weights or anything. you can view the predictions vs the target weather in simple plots.